+++ src/main/java/com/github/alexdlaird/ngrok/NgrokClient.java
@@ -181,7 +181,7 @@ public class NgrokClient {
 
     private static final Logger LOGGER = Logger.getLogger(String.valueOf(NgrokClient.class));
 
-    private static final String VERSION = "2.2.2";
+    private static final String VERSION = "2.2.3";
 
     private final JavaNgrokConfig javaNgrokConfig;
     private final NgrokProcess ngrokProcess;