+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/alexdlaird/java-ngrok/compare/2.2.2...HEAD)
+### Added
+- Documentation improvements.
 
 ## [2.2.2](https://github.com/alexdlaird/java-ngrok/compare/2.2.1...2.2.2) - TBD
 ### Added